Installer adjustments

+++ src/Field/Command/Handler/UnassignFieldHandler.php
@@ -65,6 +65,8 @@ class UnassignFieldHandler
         $stream = $this->streams->findByNamespaceAndSlug($command->getNamespace(), $command->getStream());
         $field  = $this->fields->findByNamespaceAndSlug($command->getNamespace(), $command->getField());
 
-        return $this->assignments->delete($stream->getKey(), $field->getKey());
+        if ($stream && $field) {
+            $this->assignments->delete($stream->getKey(), $field->getKey());
+        }
     }
 }
